audit: fix glm_ocr_pro - move to OCR folder, fix API endpoint bugs, add docstrings, add demo, fix README path
