+++ OCR/glm_ocr_pro/.gitignore
@@ -1,0 +1,4 @@
+.idea/
+*.iml
+__pycache__/
+*.pyc
