--- multimodal/glm_ocr_pro/.gitignore
@@ -1,4 +1,0 @@
-.idea/
-*.iml
-__pycache__/
-*.pyc
