--- multimodal/glm_ocr_pro/requirements.txt
@@ -1,3 +1,0 @@
-streamlit>=1.33.0
-pypdfium2>=4.30.0
-pillow>=10.0.0
